+++ packages/@expo/cli/src/start/server/middleware/__tests__/InterstitialPageMiddleware-test.ts
@@ -2,7 +2,6 @@ import { getConfig, getNameFromConfig } from '@expo/config';
 import { getRuntimeVersionNullableAsync } from '@expo/config-plugins/build/utils/Updates';
 import { vol } from 'memfs';
 
-import { asMock } from '../../../../__tests__/asMock';
 import { InterstitialPageMiddleware } from '../InterstitialPageMiddleware';
 import { ServerRequest, ServerResponse } from '../server.types';
 
@@ -58,8 +57,8 @@ describe('shouldHandleRequest', () => {
 
 describe('_getProjectOptionsAsync', () => {
   it('returns the project settings from the config', async () => {
-    asMock(getNameFromConfig).mockReturnValueOnce({ appName: 'my-app' });
-    asMock(getRuntimeVersionNullableAsync).mockResolvedValueOnce('123');
+    jest.mocked(getNameFromConfig).mockReturnValueOnce({ appName: 'my-app' });
+    jest.mocked(getRuntimeVersionNullableAsync).mockResolvedValueOnce('123');
 
     const middleware = new InterstitialPageMiddleware('/');
 
@@ -78,8 +77,8 @@ describe('_getProjectOptionsAsync', () => {
     );
   });
   it('returns the project settings from the config with SDK version', async () => {
-    asMock(getNameFromConfig).mockReturnValueOnce({ appName: 'my-app' });
-    asMock(getRuntimeVersionNullableAsync).mockResolvedValueOnce(null);
+    jest.mocked(getNameFromConfig).mockReturnValueOnce({ appName: 'my-app' });
+    jest.mocked(getRuntimeVersionNullableAsync).mockResolvedValueOnce(null);
 
     const middleware = new InterstitialPageMiddleware('/');